Extract command generation
