+++ lib/bluetooth.send.js
@@ -5,7 +5,7 @@ const Promise = require('bluebird');
  * Based on Noble nodejs library.
  * Uses Noble peripherals, services, characteristics to communicate with BLE devices.
  */
-module.exports = function (device) {
+module.exports = function (device, command) {
   return new Promise((resolve, reject) => {
     if (!device.characteristics || device.characteristics.length === 0) {
       device.peripheral.disconnect();
@@ -15,7 +15,7 @@ module.exports = function (device) {
       return Promise.reject('Characteristic ' + characteristic.uuid + ' not writable for ' + device.peripheral.address);
     } else {
       const connectTimeout = setTimeout(timeout, shared.sendTimeout, reject, device.peripheral);
-      device.characteristics[0].write(new Buffer(device.command), false, (error) => {
+      device.characteristics[0].write(new Buffer(command), false, (error) => {
         connectTimeout.unref();
         device.peripheral.disconnect();
